+++ NuGet/linq2db.Informix.nuspec
@@ -26,5 +26,8 @@
 		<file src="..\Redist\IBM\IBM.Data.Informix.dll" target="tools" />
 		<file src="LinqToDB.Informix.Tools.ttinclude"   target="content\LinqToDB.Templates" />
 		<file src="CopyMe.Informix.tt.txt"              target="content\LinqToDB.Templates" />
+
+		<file src="init.ps1"                            target="tools\init.ps1" />
+
 	</files>
 </package>
