+++ NuGet/linq2db.MySql.nuspec
@@ -25,5 +25,8 @@
 		<file src="..\packages\MySql.Data.6.9.9\lib\net40\MySql.Data.dll" target="tools" />
 		<file src="LinqToDB.MySql.Tools.ttinclude"                        target="content\LinqToDB.Templates" />
 		<file src="CopyMe.MySql.tt.txt"                                   target="content\LinqToDB.Templates" />
+
+		<file src="init.ps1"                                              target="tools\init.ps1" />
+		<file src="CopyMe.MySql.Core.tt.txt"                              target="tools" />
 	</files>
 </package>
