+++ NuGet/linq2db.t4models.nuspec
@@ -18,36 +18,30 @@
 		</tags>
 
 		<contentFiles>
-			<files include="**/content/**/*.*" buildAction="None" />
+			<files include="**/*.*" buildAction="None" />
 		</contentFiles>
+		<dependencies>
+			
+			<group targetFramework="netstandard1.6">
+				<dependency id="linq2db.Core" version="1.8.0"/>
+			</group>
+			
+			<group targetFramework="net40">
+				<dependency id="linq2db" version="1.8.0"/>
+			</group>
+
+		</dependencies>
 	</metadata>
 	<files>
 		<file src="..\packages\linq2db.1.8.0\lib\net40\linq2db.dll"    target="tools" />
 		<file src="init.ps1"                                           target="tools\init.ps1" />
 
 		<file src="..\Templates\ReadMe.LinqToDB.md"                    target="content\LinqToDB.Templates" />
-		<file src="..\Templates\DataAnnotations.ttinclude"             target="content\LinqToDB.Templates" />
-		<file src="..\Templates\DataAnnotations.ttinclude"             target="content\LinqToDB.Templates" />
-		<file src="..\Templates\DataModel.ttinclude"                   target="content\LinqToDB.Templates" />
-		<file src="..\Templates\EditableObject.ttinclude"              target="content\LinqToDB.Templates" />
-		<file src="..\Templates\ObsoleteAttributes.ttinclude"          target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.Access.ttinclude"             target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.Firebird.ttinclude"           target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.MySql.ttinclude"              target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.Oracle.ttinclude"             target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.PostgreSQL.ttinclude"         target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.SqlCe.ttinclude"              target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.SQLite.ttinclude"             target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.SqlServer.ttinclude"          target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.SqlServer.SqlTypes.ttinclude" target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.Sybase.ttinclude"             target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.DB2.ttinclude"                target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.Informix.ttinclude"           target="content\LinqToDB.Templates" />
-		<file src="..\Templates\LinqToDB.ttinclude"                    target="content\LinqToDB.Templates" />
-		<file src="..\Templates\NotifyPropertyChanged.ttinclude"       target="content\LinqToDB.Templates" />
-		<file src="..\Templates\PluralizationService.ttinclude"        target="content\LinqToDB.Templates" />
-		<file src="..\Templates\T4Model.ttinclude"                     target="content\LinqToDB.Templates" />
-		<file src="..\Templates\Validation.ttinclude"                  target="content\LinqToDB.Templates" />
+		<file src="..\Templates\*.ttinclude"                           target="content\LinqToDB.Templates" />
 		<file src="LinqToDB.Tools.ttinclude"                           target="content\LinqToDB.Templates" />
+
+		<file src="..\Templates\ReadMe.LinqToDB.md"                    target="contentFiles\LinqToDB.Templates" />
+		<file src="..\Templates\*.ttinclude"                           target="contentFiles\LinqToDB.Templates" />
+		<file src="LinqToDB.Tools.ttinclude"                           target="contentFiles\LinqToDB.Templates" />
 	</files>
 </package>
