+++ T4Models.2017.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26228.12
+VisualStudioVersion = 15.0.26430.6
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Templates", "Templates", "{2E57C99E-2603-4D59-9094-A5BB06639C10}"
 	ProjectSection(SolutionItems) = preProject
@@ -34,20 +34,26 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "NuGet", "NuGet", "{2AA0D9F0
 	ProjectSection(SolutionItems) = preProject
 		NuGet\CopyMe.Access.tt.txt = NuGet\CopyMe.Access.tt.txt
 		NuGet\CopyMe.DB2.tt.txt = NuGet\CopyMe.DB2.tt.txt
+		NuGet\CopyMe.Firebird.Core.tt.txt = NuGet\CopyMe.Firebird.Core.tt.txt
 		NuGet\CopyMe.Firebird.tt.txt = NuGet\CopyMe.Firebird.tt.txt
 		NuGet\CopyMe.Informix.tt.txt = NuGet\CopyMe.Informix.tt.txt
+		NuGet\CopyMe.MySql.Core.tt.txt = NuGet\CopyMe.MySql.Core.tt.txt
 		NuGet\CopyMe.MySql.tt.txt = NuGet\CopyMe.MySql.tt.txt
 		NuGet\CopyMe.Oracle.Managed.tt.txt = NuGet\CopyMe.Oracle.Managed.tt.txt
 		NuGet\CopyMe.Oracle.x64.tt.txt = NuGet\CopyMe.Oracle.x64.tt.txt
 		NuGet\CopyMe.Oracle.x86.tt.txt = NuGet\CopyMe.Oracle.x86.tt.txt
+		NuGet\CopyMe.PostgreSQL.Core.tt.txt = NuGet\CopyMe.PostgreSQL.Core.tt.txt
 		NuGet\CopyMe.PostgreSQL.tt.txt = NuGet\CopyMe.PostgreSQL.tt.txt
 		NuGet\CopyMe.SapHana.tt.txt = NuGet\CopyMe.SapHana.tt.txt
 		NuGet\CopyMe.SqlCe.tt.txt = NuGet\CopyMe.SqlCe.tt.txt
+		NuGet\CopyMe.SQLite.Core.tt.txt = NuGet\CopyMe.SQLite.Core.tt.txt
 		NuGet\CopyMe.SQLite.tt.txt = NuGet\CopyMe.SQLite.tt.txt
+		NuGet\CopyMe.SqlServer.Core.tt.txt = NuGet\CopyMe.SqlServer.Core.tt.txt
 		NuGet\CopyMe.SqlServer.tt.txt = NuGet\CopyMe.SqlServer.tt.txt
 		NuGet\CopyMe.Sybase.tt.txt = NuGet\CopyMe.Sybase.tt.txt
 		NuGet\CopyMe.tt.txt = NuGet\CopyMe.tt.txt
 		NuGet\Init.ps1 = NuGet\Init.ps1
+		NuGet\install.ps1 = NuGet\install.ps1
 		NuGet\linq2db.Access.nuspec = NuGet\linq2db.Access.nuspec
 		NuGet\linq2db.DB2.nuspec = NuGet\linq2db.DB2.nuspec
 		NuGet\linq2db.Firebird.nuspec = NuGet\linq2db.Firebird.nuspec
