+++ Tests/LinqToDB/packages.config
@@ -3,7 +3,7 @@
   <package id="FirebirdSql.Data.FirebirdClient" version="5.6.0" targetFramework="net45" />
   <package id="linq2db" version="1.8.0" targetFramework="net45" />
   <package id="Microsoft.SqlServer.Compact" version="4.0.8876.1" targetFramework="net40" />
-  <package id="Microsoft.SqlServer.Types" version="11.0.2" targetFramework="net45" />
+  <package id="Microsoft.SqlServer.Types" version="14.0.314.76" targetFramework="net45" />
   <package id="MySql.Data" version="6.9.9" targetFramework="net45" />
   <package id="Npgsql" version="2.2.6" targetFramework="net45" />
   <package id="Oracle.ManagedDataAccess" version="12.1.24160719" targetFramework="net45" />
